refactor(chat): separate palette state and view
